fix stack too deep error

+++ lib/devise-security/models/password_archivable.rb
@@ -61,9 +61,12 @@ module Devise
 
       private
 
-      # archive the last password before save and delete all to old passwords from archive
+      # Archive the last password before save and delete all to old passwords from archive
+      # @note we check to see if an old password has already been archived because
+      #   mongoid will keep re-triggering this callback when we add an old password
       def archive_password
         if max_old_passwords > 0
+          return if old_passwords.where(encrypted_password: encrypted_password_was).exists?
           old_passwords.create!(encrypted_password: encrypted_password_was) if encrypted_password_was.present?
           old_passwords.order(created_at: :desc).offset(max_old_passwords).destroy_all
         else
